Moved feedback code in its own file, added some props, regular cn usage
