+++ front/lib/resources/agent_message_feedback_resource.ts
@@ -125,9 +125,9 @@ export class AgentMessageFeedbackResource extends BaseResource<AgentMessageFeedb
     paginationParams: PaginationParams;
   }) {
     const where: WhereOptions<AgentMessageFeedback> = {
-      // IMPORTANT: Necessary for global models who share ids across workspaces.
+      // Safety check: global models share ids across workspaces and some have had feedbacks.
       workspaceId: workspace.id,
-      agentConfigurationId: agentConfiguration.sId.toString(),
+      agentConfigurationId: agentConfiguration.sId,
     };
 
     // Get the total feedback count, because needed for pagination
